+++ packages/metalsmith-include-files/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v2.0.0 / 2023-10-31
+
+- Converted to be an ES6 module.
+
 ## v1.1.0 / 2023-10-30
 
 - Added `Metalsmith#debug()` calls.